Extract Mattermost thread backfill helpers
